+++ src/raters/font-awesome/FaBaseGlyph.vue
@@ -1,8 +1,7 @@
 <script type="text/javascript">
-import FaBaseGlyph from './FaBaseGlyph.vue';
-import Glyphs from './glyphs.js';
+import Path from '../../templates/Path.vue';
 
-export default FaBaseGlyph.extend({
+export default Path.extend({
     props: {
         customProps: {
             required: true,
@@ -10,13 +9,7 @@ export default FaBaseGlyph.extend({
         }
     },
     created() {
-        this.updateGlyph();
-    },
-    methods: {
-        updateGlyph() {
-            let glyph = this.customProps.glyph.replace(/^fa\-/, "");
-            this.points = [Glyphs[glyph]];
-        }
+      this.coords.x1="-2%";
     },
     data() {
         return {