+++ changelog.md
@@ -1,8 +1,14 @@
 # Change Log
 - Project follows [Semantic Versioning](http://semver.org/)
 
+## [Unreleased]
+- Added export of individual font awesome glyphs
+- Added extendable base glyph for font-awesome to allow use of specific glyph
+
 ## Version 1.0 (2017-04-26)
 - Initial release
+
+